requirements.txt
pandas==2.2.2
requests==2.32.3
pyyaml==6.0.2
